refactor(is/focus-relevant): allow running identification with partial exceptions
This commit is a follow-up to
bb69ccb fix(is/focus-relevant): identify flexbox containers in IE
